requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-docx==1.1.0
PyPDF2==3.0.1
openai==1.6.1
langchain==0.0.350
langchain-openai==0.0.2
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.3
plotly==5.17.0
jinja2==3.1.2
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.12.1
reportlab==4.0.7
email-validator==2.1.0
